backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
boto3
python-multipart
python-dotenv
email-validator
setuptools
psycopg2